pyproject.toml
[tool.poetry]
name = "crypto_aggregator"
version = "0.1.0"
description = ""
authors = ["Simon Yerro"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
uvicorn = "^0.15.0"
python-dotenv = "^0.19.1"
fastapi = "^0.70.0"
firebase-admin = "^5.0.3"

[tool.poetry.dev-dependencies]
ipykernel = "^6.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
